Write a modern and better to test scrapy import plugin which uses scrapy selectors.
The scrapy selectors are easy to use while parsing websites. The essen und
trinken page was implemented as a first test. Furthermore there is a test-case
attached which enables us to create reference data from the webpage to ensure
the plugins will stay working. They use a set of reference data which can
be produced also by running the testcase itself. This makes it easy to
maintain because only the outcome has to be checked.